Change value from Int to Double

+++ Tablut/src/it/unibo/ai/didattica/competition/tablut/client/player/heuristic/util/HeuristicElement.kt
@@ -1,3 +1,3 @@
 package it.unibo.ai.didattica.competition.tablut.client.player.heuristic.util
 
-class HeuristicElement(val name: String, val value: Int, val min: Int, val max: Int, val factor: Double) {}
+class HeuristicElement(val name: String, val value: Double, val min: Int, val max: Int, val factor: Double) {}
